Split HAP server module and centralize deploy swap helpers.
Readability-only hap/server/ split; Channel::next() replaces duplicated
driver cycling; upgrade binary swap/rollback uses shared deploy helpers.
